fix: move loss to model.py
